move ProviderMetaInfo to model package
